Fix KITTI conversion minor bug
- When running KITTI conversion script, the script attempted to convert
a directory of training data that did not exist. I removed this
directory from the list of directories to attempt conversion for.
- This could be made more general by using `os.listdir` or whatever the
function is to iterate through a directory, but this seems sufficient.
- Lastly, formats the Python file according to my formatter rules I
typically use. I'm not a stickler for this, it just happened
automatically and I didn't think it would hurt to keep it.